Restructured core.astro_object
