+++ src/core/astro_object/infrastructure.py
@@ -1,13 +1,9 @@
 from returns.result import Success, Failure
 from db_plugins.db.mongo import models
 
-from shared.error.exceptions import (
-    ServerErrorException,
-    ClientErrorException,
-    EmptyQuery
-)
+from shared.error.exceptions import ClientErrorException, EmptyQuery
 from shared.utils.repositories import MongoRepository
-from ..domain.astro_object_service import AstroObjectPayload
+from .payload import AstroObjectPayload
 
 
 class ListAstroObjectRepository(MongoRepository):
@@ -21,5 +17,19 @@ class ListAstroObjectRepository(MongoRepository):
     def _wrap_results(self, result):
         if result.total > 0:
             return Success(list(result))
+        else:
+            return Failure(ClientErrorException(EmptyQuery()))
+
+
+class SingleAstroObjectRepository(MongoRepository):
+    def _query(self, payload: AstroObjectPayload):
+        return self.db.query().find_one(
+            model=models.Object,
+            filter_by=payload.filter_by
+        )
+
+    def _wrap_results(self, result):
+        if result:
+            return Success(result)
         else:
             return Failure(ClientErrorException(EmptyQuery()))
