+++ src/core/astro_object/payload.py
@@ -12,8 +12,8 @@ class AstroObjectPayload(PayloadFactory):
     _rules = {
         'aid': FilterRules(
             ['aid'],
-            None,
-            str
+            '$in',
+            AstroObjectHelpers.list_of_str
         ),
         'oid': FilterRules(
             ['oid'],
@@ -41,15 +41,3 @@ class AstroObjectPayload(PayloadFactory):
             AstroObjectHelpers.query_for_locs
         )
     }
-
-
-class AstroObjectService:
-    def __init__(self, list_object_repository, single_object_repository):
-        self.list_object_repository = list_object_repository
-        self.single_object_repository = single_object_repository
-
-    def get_list_object(self, payload: AstroObjectPayload):
-        return self.list_object_repository.get(payload)
-
-    def get_single_object(self, payload: AstroObjectPayload):
-        return self.single_object_repository.get(payload)
