--- src/core/astro_object/use_case/get_single_object.py
@@ -1,13 +1,0 @@
-from shared.interface.command import Command, ResultHandler
-from core.astro_object.domain.astro_object_service import AstroObjectService
-
-
-class GetSingleAstroObject(Command):
-    def __init__(
-            self,
-            service,
-            payload: AstroObjectService,
-            handler: ResultHandler
-    ):
-        super().__init__(service, payload, handler)
-        self.action = 'get_single_object'
